Split useLabeler.ts into hooks/labeler/ modules
Pure restructure, no behavior change - useLabeler.ts drops from ~1400
to ~830 lines and stays the public entrypoint (all external imports
of @renderer/hooks/useLabeler are unaffected via re-exports):

- labeler/storeTypes.ts: Rect/Vector2 and the store's state/action/
  history/hit-test types
- labeler/hitIds.ts: OneToOneMap + the color-hit-id bookkeeping
  (createHitIdTracker) used by canvas hit-testing
- labeler/geometry.ts: pure point/shape math - normalizeAnnotationPoints,
  axisMaskedMoveDelta, box<->polygon conversion, duplicate offsetting,
  isValidBoxCreation (now a pure function, taking rawStart explicitly
  instead of closing over module state)
- labeler/history.ts: undo/redo stack + the optimistic create/delete/
  replace-points/relabel/convert-type "apply*" primitives

useLabeler.ts keeps the zustand create() call and the actions that are
inherently store-shaped (mode/mouse/selection/box-creation flow, zoom,
viewport math).
